Move CreateRemoteThread code to util method
